+++ src/mkvtoolnix-gui/util/settings.h
@@ -148,6 +148,7 @@ public:
   using RunProgramConfigPtr  = std::shared_ptr<RunProgramConfig>;
   using RunProgramConfigList = QList<RunProgramConfigPtr>;
 
+  bool m_useLegacyFontMIMETypes;
   mtx::bcp47::language_c m_defaultAudioTrackLanguage, m_defaultVideoTrackLanguage, m_defaultSubtitleTrackLanguage, m_defaultChapterLanguage;
   SetDefaultLanguagePolicy m_whenToSetDefaultLanguage;
   DeriveLanguageFromFileNamePolicy m_deriveAudioTrackLanguageFromFileNamePolicy, m_deriveVideoTrackLanguageFromFileNamePolicy, m_deriveSubtitleTrackLanguageFromFileNamePolicy;